All pages are loading correctly and only css lacks
