+++ index.html
@@ -12,6 +12,7 @@
     <link rel="stylesheet" href="styles/meny-styles-non-responsive.css" />
     <link rel="stylesheet" href="styles/pizza-management.css" />
     <link rel="stylesheet" href="styles/styles.css">
+    <link rel="stylesheet" href="styles/orderConfirmation.css">
     <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;500&display=swap" rel="stylesheet">
 
 
@@ -53,20 +54,17 @@
         <li id="signIn-id">
           <a href="/signIn" data-navigo>Tilmeld</a>
         </li>
-        <li id="addPizza-id" style="display: none;">
-          <a href="/addPizza" data-navigo>Tilføj Pizza</a>
-        </li>  
+          <li id="ingredients-id" style="display: none;">
+              <a href="/ingredients" data-navigo>Ingredienser</a>
+          </li>
+        <li id="pizza-management-id" style="display: none;">
+          <a href="/pizzaBehandling" data-navigo>Pizza Behandling</a>
+        </li>
         <li id="login-id">
           <a href="/login" data-navigo>Login</a>
         </li>
         <li id="logout-id" style="display:none">
           <a href="/logout" data-navigo>Logout</a>
-        </li>
-          <li id="ingredients-id">
-              <a href="/ingredients" data-navigo>Ingredienser</a>
-          </li>
-        <li id="pizza-management-id">
-          <a href="/pizzaBehandling" data-navigo>Pizza Behandling</a>
         </li>
       </ul>
     </nav>
