--- pages/addPizzasPage/addPizzasPage.html
@@ -1,6 +1,0 @@
-<template>
-    <h2>Tilføj Pizza</h2>
-    <p>Her kan der komme en masse info "had id been a real project"</p>
-    <p>See <a href="https://getbootstrap.com/docs/5.0/components/navbar/" target="_blank">here</a> for details about the
-      responsive Bootstrap navbar used in this example</p>
-<template>
