--- pages/addPizzasPage/addPizzasPage.js
@@ -1,4 +1,0 @@
-console.log("loading order")
-export function initAddPizza(){
-    console.log("Hello from Adding pizzas!")
-}
