+++ pages/orderConfirmation/orderConfirmation.html
@@ -1,6 +1,28 @@
-<template>
-    <h2>Ubekræftede ordre</h2>
-    <p>Her kan der komme en masse info "had id been a real project"</p>
-    <p>See <a href="https://getbootstrap.com/docs/5.0/components/navbar/" target="_blank">here</a> for details about the
-      responsive Bootstrap navbar used in this example</p>
-<template>
+  <template>
+  <div class="container">
+    <h1>Order Confirmation</h1>
+
+    <div class="order-wrapper">
+      <div class="order-info" id="order-info">
+        <!-- Order information will be populated by JavaScript -->
+      </div>
+
+      <div class="time-options">
+        <h2>Choose Pickup Time</h2>
+        <button class="pickup-time" onclick="setPickupTime(10, this)">10 Minutes</button>
+        <button class="pickup-time" onclick="setPickupTime(15, this)">15 Minutes</button>
+        <button class="pickup-time" onclick="setPickupTime(20, this)">20 Minutes</button>
+        <button class="pickup-time" onclick="setPickupTime(30, this)">30 Minutes</button>
+        <button class="pickup-time" onclick="setPickupTime(45, this)">45 Minutes</button>
+        <button class="pickup-time" onclick="setPickupTime(60, this)">60 Minutes</button>
+        <button class="pickup-time" onclick="setPickupTime(90, this)">90 Minutes</button>
+        <!-- ... rest of your buttons ... -->
+      </div>
+      
+
+    <div class="confirmation-buttons">
+      <button class="confirm-button" onclick="confirmOrder()">Confirm Order</button>
+      <button class="refuse-button" onclick="refuseOrder()">Refuse Order</button>
+    </div>
+  </div>
+</template>
