+++ pages/pizzaManagement/pizzaManagement.html
@@ -48,7 +48,7 @@
             </div>
             <div class="col-md-5">
                 <div class="mt-4">
-                    <h2>Nuværrende Pizzaer</h2>
+                    <h2>Nuværende Pizzaer</h2>
                     <input type="text" id="searchInput" placeholder="Søg efter navn">
                     <div id="pizzaContainer" class="pizza-management-pizza-container">
                         <ul id="pizzaManagement-pizzaList"></ul>
