+++ styles/orderConfirmation.css
@@ -1,13 +1,14 @@
 /* Styles for Order Confirmation page */
 
 /* Global styles */
-
-{
+/*
+  {
     box-sizing: border-box;
     font-family: Arial, sans-serif;
     margin: 0;
     padding: 0;
   }
+  */
   body {
     background-color: #f4f4f4;
   }
